fix: move trigger_parser into teatree package, replace scripts/lib copy with symlink
_startup.py imported trigger_parser via sys.path hack from scripts/lib/.
When teatree is installed from git (not editable), scripts/lib/ doesn't
exist, causing ModuleNotFoundError in overlay CI.

Refactored parse_triggers to comply with C901/PLR0912 complexity limits.
